Completed TemplateSword
TemplateSword.cs ks finished and moved to TemplateMod.Items.Weapons

Maybe update TemplateSword.png
