--- Items/TemplateSword.cs
@@ -1,38 +1,0 @@
-using Terraria.ID;
-using Terraria.ModLoader;
-
-namespace TemplateMod.Items
-{
-	public class TemplateSword : ModItem
-	{
-		public override void SetStaticDefaults()
-		{
-			DisplayName.SetDefault("TemplateSword");
-			Tooltip.SetDefault("This is a modded sword.");
-		}
-		public override void SetDefaults()
-		{
-			item.damage = 50;
-			item.melee = true;
-			item.width = 40;
-			item.height = 40;
-			item.useTime = 20;
-			item.useAnimation = 20;
-			item.useStyle = 1;
-			item.knockBack = 6;
-			item.value = 10000;
-			item.rare = 2;
-			item.UseSound = SoundID.Item1;
-			item.autoReuse = true;
-		}
-
-		public override void AddRecipes()
-		{
-			ModRecipe recipe = new ModRecipe(mod);
-			recipe.AddIngredient(ItemID.DirtBlock, 10);
-			recipe.AddTile(TileID.WorkBenches);
-			recipe.SetResult(this);
-			recipe.AddRecipe();
-		}
-	}
-}
