+++ Items/Weapons/TemplateSword.png
